@@ -2,14 +2,15 @@
   "name": "board4you",
   "version": "1.0.6",
   "description": "client side of board4you",
-  "main": "./src/main/main.js",
+  "main": "./public/main.js",
   "scripts": {
-    "build": "webpack build --mode=production --no-devtool",
-    "dev": "webpack -w --mode=development",
+    "build": "./node_modules/.bin/esbuild ./src/entries/web.tsx --bundle --outdir=./public --loader:.png=dataurl --minify",
+    "dev": "./node_modules/.bin/esbuild ./src/entries/web.tsx --bundle --outdir=./public --loader:.png=dataurl --watch",
+    "buildMain": "./node_modules/.bin/esbuild ./src/entries/main.ts --bundle --outdir=./public --loader:.png=dataurl --platform=node",
+    "devMain": "./node_modules/.bin/esbuild ./src/entries/main.ts --bundle --outdir=./public --loader:.png=dataurl --platform=node --watch",
     "electron": "electronmon .",
-    "start": "webpack serve --open --mode=development",
-    "buildApp": "webpack build --mode=production --no-devtool; electron-builder -wl",
-    "release": "webpack build --mode=production --no-devtool; electron-builder -wl",
+    "buildApp": "./node_modules/.bin/esbuild ./src/entries/web.tsx --bundle --outdir=./public --loader:.png=dataurl --minify; electron-builder -wl",
+    "release": "./node_modules/.bin/esbuild ./src/entries/web.tsx --bundle --outdir=./public --loader:.png=dataurl --minify; electron-builder -wl",
     "test": "jest"
   },
   "build": {
@@ -53,10 +54,10 @@
     "bootstrap": "^5.2.3",
     "electron-updater": "^5.3.0",
     "fbemitter": "^3.0.0",
-    "image-to-pdf": "^2.0.0",
     "is-url": "^1.2.4",
     "jspdf": "^2.5.1",
     "konva": "^8.3.14",
+    "pdfkit": "^0.13.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.6.0",
     "react-bootstrap-range-slider": "^3.0.8",
@@ -75,28 +76,22 @@
     "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.5",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.22.5",
     "@faker-js/faker": "^8.0.2",
     "@types/fbemitter": "^2.0.32",
-    "@types/file-loader": "^5.0.2",
     "@types/jest": "^29.5.3",
+    "@types/pdfkit": "^0.13.1",
     "@types/react-color": "^3.0.6",
     "@types/react-dom": "^18.2.4",
     "@types/react-router-bootstrap": "^0.26.3",
     "@types/uuid": "^9.0.1",
     "@typescript-eslint/eslint-plugin": "^5.59.8",
     "@typescript-eslint/parser": "^5.59.8",
-    "babel-loader": "^9.1.0",
     "css-loader": "^6.7.3",
-    "electron": "^23.2.0",
+    "electron": "26.4.1",
     "electron-builder": "^22.10.3",
     "electron-rebuild": "^3.2.9",
     "electronmon": "^2.0.2",
+    "esbuild": "^0.19.5",
     "eslint": "^8.41.0",
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-import": "^2.27.5",
@@ -105,18 +100,11 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.0",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.6.2",
     "pdfjs-dist": "^3.7.107",
     "prop-types": "^15.8.1",
     "react-router-dom": "^6.16.0",
-    "style-loader": "^3.3.1",
-    "ts-loader": "^9.4.3",
-    "typescript": "^5.2",
-    "webpack": "^5.75.0",
-    "webpack-bundle-analyzer": "^4.8.0",
-    "webpack-cli": "^5.0.0",
-    "webpack-dev-server": "^4.11.1"
+    "typescript": "^5.2"
   }
 }
